Extract the validator methods from Attributes::Dsl
